+++ scripts/changelog.ts
@@ -200,6 +200,41 @@ async function exec(): Promise<void> {
         ],
       ],
     },
+    {
+      tagName: 'v0.7.0',
+      name: '0.7.0',
+      date: '2020-08-04',
+      breakingChanges: [
+        [
+          'notification',
+          'Completely new implementation. We now basically resolve the different event types in the API directly. See `AbstractNotificationEvent`.',
+        ],
+        [
+          'uuid',
+          'Rename `article` to `repository` in `ArticleRevision` (and similarly for all other types that implement `AbstractRevision`.',
+        ],
+      ],
+      added: [
+        [
+          'uuid',
+          `Add various base types that expose common properties. More specifically:
+- Add base type \`AbstractExercise\` for \`Exercise\` and \`GroupedExercise\`.
+- Add base type \`AbstractExerciseRevision\` for \`ExerciseRevision\` and \`GroupedExerciseRevision\`.
+- Add base type \`AbstractNavigationChild\` for \`Page\` and \`TaxonomyTerm\`.
+- Add base type \`AbstractRepository\` for \`AbstractEntity\` and \`Page\`.
+- Add base type \`AbstractRepositoryRevision\` for \`AbstractEntityRevision\` and \`PageRevision\`.
+- Add base type \`AbstractTaxonomyTermChild\` for entity types that appear as children of taxonomy terms.
+              `,
+        ],
+        ['uuid', 'Add field `date` to `Page`.'],
+      ],
+      changed: [
+        [
+          'uuid',
+          '`AbstractEntity`, `AbstractEntityRevision` inherit the properties of `AbstractUuid` explicitly',
+        ],
+      ],
+    },
   ])
 
   await writeFile(path.join(__dirname, '..', 'CHANGELOG.md'), content)